+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## Version 0.0.5
+## Version 0.1.0
 
-* Run transmission as a less privilege user
+* Run transmission as a less privileged user
 * Transmission RPC authentication enabled by default
-* Firewall rules for the traffic
+* Optional firewall for stricter rules
+* Protocol wise VPN config split to support providers with both TCP and UDP
+* Support FastestVPN VPN provider
 
 ## Version 0.0.4